rename command to bulk
